# Split state variables and symbolic variables. Stricter type checking.