--- a/backends/p4tools/common/core/solver.h
+++ b/backends/p4tools/common/core/solver.h
@@ -4,7 +4,11 @@
 #include <optional>
 #include <vector>
 
-#include "backends/p4tools/common/lib/model.h"
+#include <boost/container/flat_map.hpp>
+#include <boost/container/flat_set.hpp>
+
+#include "ir/ir.h"
+#include "lib/castable.h"
 #include "lib/cstring.h"
 
 namespace P4Tools {
@@ -12,6 +16,18 @@ namespace P4Tools {
 /// Represents a constraint that can be shipped to and asserted within a solver.
 // TODO: This should implement AbstractRepCheckedNode<Constraint>.
 using Constraint = IR::Expression;
+
+struct SymbolicVarComp {
+    bool operator()(const IR::SymbolicVariable *s1, const IR::SymbolicVariable *s2) const {
+        return s1->operator<(*s2);
+    }
+};
+
+/// This type maps symbolic variables to their value assigned by the solver.
+using SymbolicMapping = boost::container::flat_map<const IR::SymbolicVariable *,
+                                                   const IR::Expression *, SymbolicVarComp>;
+
+using SymbolicSet = boost::container::flat_set<const IR::SymbolicVariable *, SymbolicVarComp>;
 
 /// Provides a higher-level interface for an SMT solver.
 class AbstractSolver : public ICastable {
@@ -40,13 +56,13 @@ class AbstractSolver : public ICastable {
     /// @checkSat returned anything other than true, if there was no such previous call, or if the
     /// state in the solver has changed since the last such call (e.g., more assertions have been
     /// made).
-    virtual const Model *getModel() const = 0;
+    [[nodiscard]] virtual const SymbolicMapping &getSymbolicMapping() const = 0;
 
     /// Saves solver state to the given JSON generator.
     virtual void toJSON(JSONGenerator &) const = 0;
 
     /// @returns whether this solver is incremental.
-    virtual bool isInIncrementalMode() const = 0;
+    [[nodiscard]] virtual bool isInIncrementalMode() const = 0;
 
     /// Cast an abstract solver to its specific sub type.
     /// Casts involving "const" have to explicit in this case. This is necessary because abstract