--- a/backends/p4tools/common/core/solver.h
+++ b/backends/p4tools/common/core/solver.h
@@ -17,6 +17,7 @@ namespace P4Tools {
 // TODO: This should implement AbstractRepCheckedNode<Constraint>.
 using Constraint = IR::Expression;
 
+/// Comparator to compare SymbolicVariable pointers.
 struct SymbolicVarComp {
     bool operator()(const IR::SymbolicVariable *s1, const IR::SymbolicVariable *s2) const {
         return s1->operator<(*s2);