--- a/backends/p4tools/common/lib/variables.cpp
+++ b/backends/p4tools/common/lib/variables.cpp
@@ -2,7 +2,7 @@
 
 #include <map>
 #include <string>
-#include <utility>
+#include <tuple>
 
 #include "ir/id.h"